requirements.txt
requests
beautifulsoup4
lxml
google-generativeai